Extract store configuration from module
