+++ ui/src/app/cohort-search/cohort-search.module.ts
@@ -2,8 +2,7 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {ClarityModule} from 'clarity-angular';
-import {NgReduxModule, NgRedux, DevToolsExtension} from '@angular-redux/store';
-import {createEpicMiddleware, combineEpics} from 'redux-observable';
+import {NgReduxModule} from '@angular-redux/store';
 
 /* Components */
 import {CohortBuilderComponent} from './cohort-builder/cohort-builder.component';
@@ -21,16 +20,11 @@ import {
 
 /* Other Objects */
 import {CohortSearchRouter} from './router.module';
-import {environment} from 'environments/environment';
-
 import {
   CohortSearchActions,
   CohortSearchEpics,
-  CohortSearchState,
-  initialState,
-  rootReducer
+  ConfigureStore,
 } from './redux';
-
 import {CohortBuilderService} from 'generated';
 
 // tslint:enable:max-line-length
@@ -61,32 +55,9 @@ import {CohortBuilderService} from 'generated';
     CohortBuilderService,
     CohortSearchActions,
     CohortSearchEpics,
+    ConfigureStore,
   ]
 })
 export class CohortSearchModule {
-  constructor(private ngRedux: NgRedux<CohortSearchState>,
-              private epics: CohortSearchEpics,
-              private devTools: DevToolsExtension) {
-
-    let storeEnhancers = [];
-    if (environment.debug && devTools.isEnabled()) {
-      storeEnhancers = [...storeEnhancers, devTools.enhancer()];
-    }
-
-    const middleware = [
-      createEpicMiddleware(
-        combineEpics(
-          this.epics.fetchCriteria,
-          this.epics.fetchCounts,
-        )
-      )
-    ];
-
-    ngRedux.configureStore(
-      rootReducer,
-      initialState,
-      middleware,
-      storeEnhancers
-    );
-  }
+  constructor(store: ConfigureStore) {}
 }
