+++ ui/src/app/cohort-search/redux/store/selectors.ts
@@ -1,53 +1,8 @@
-import {Map, List, Set, fromJS} from 'immutable';
+import {Map, List} from 'immutable';
 import {Criteria, SearchRequest} from 'generated';
 
-/**
- * InitialState
- */
-export const SR_ID = 'searchReqest0';
+import {SR_ID} from './initial';
 
-export const initialState = fromJS({
-  entities: {
-    searchRequests: {
-      [SR_ID]: {
-        includes: ['include0'],
-        excludes: ['exclude0'],
-        count: 0,
-        isRequesting: false
-      }
-    },
-    groups: {
-      include0: {
-        id: 'include0',
-        items: [],
-        count: null,
-        isRequesting: false,
-      },
-      exclude0: {
-        id: 'exclude0',
-        items: [],
-        count: null,
-        isRequesting: false,
-      },
-    },
-    items: {},
-    criteria: {},
-  },
-
-  wizard: {
-    open: false,
-    item: {},
-    selections: {},
-  },
-
-  criteria: {
-    tree: {},
-    requests: {},
-  },
-});
-
-
-export type CohortSearchState = Map<any, any>;
 
 /**
  * Entities
