+++ ui/src/app/cohort-search/redux/store/types.ts
@@ -1,0 +1,3 @@
+import {Map} from 'immutable';
+
+export type CohortSearchState = Map<any, any>;
